% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_estimate_mean_treat.R
\name{esti_mean_treat}
\alias{esti_mean_treat}
\title{Estimate treatment effect and the corresponding variance estimation on the treatment arm using different covariates adjustment methods.}
\usage{
esti_mean_treat(X, Y, A, H = NULL)
}
\arguments{
\item{X}{The n by p covariates matrix.}

\item{Y}{Vector of n dimensional observed response.}

\item{A}{Vector of n dimensional treatment assignment.}

\item{H}{The n by n hat projection matrix corresponding to X.}
}
\value{
the list of point estimation and variance estimation for all kinds of estimators.
\item{tau_unadj}{The point estimation of the unadjusted estimator tau_unadj.}
\item{tau_hat_lin}{The point estimation of Lin’s (2013) adjusted estimator.}
\item{tau_hat_lin_db}{The point estimation of Lei’s (2020) debiased adjusted estimator using population leverage scores.}
\item{tau_db}{The point estimation of the debiased estimator by Lu et al.(2023).}
\item{tau_adj2c}{The point estimation of another form of the debiased estimator inspired by HOIF (Zhao et al.(2024)).}
\item{tau_adj2}{The point estimation of the adjusted estimator motivatied by HOIF (Zhao et al.(2024)).}
\item{tau_adj3}{The point estimation of the bias-free adjusted estimator based on tau_adj2.}
\item{var_hat_unadj}{The oracle variance of the unadjusted estimator tau_unadj.}
\item{var_hat_lin}{The HC3 type variance estimation of Lin’s (2013) adjusted estimator tau_hat_lin.}
\item{var_hat_lin_db}{The HC3 type variance estimation of Lei’s (2020) adjusted estimator tau_hat_lin_db.}
\item{var_hat_db}{The variance estimation of the debiased estimator tau_db.}
\item{var_hat_adj2c_wang}{The variance estimation of the adjusted estimator tau_adj2c where the estimation method follows Lu et al.(2023).}
\item{var_hat_adj2c_v2}{The variance estimation of the adjusted estimator tau_adj2c using the conservative method introduced in Zhao et al.(2024).}
\item{var_hat_adj2_wang}{The variance estimation of the adjusted estimator tau_adj2 where the estimation method follows Lu et al.(2023).}
\item{var_hat_adj2_v2}{The variance estimation of the adjusted estimator tau_adj2 using using the conservative method introduced in Zhao et al.(2024).}
}
\description{
Estimate treatment effect and the corresponding variance estimation on the treatment arm using different covariates adjustment methods.
}
\examples{
set.seed(100)
n <- 500
p <- n * 0.3
beta <- runif(p, -1 / sqrt(p), 1 / sqrt(p))

X <- mvtnorm::rmvt(n, sigma = diag(1, p), df = 3)
Y1 <- as.numeric(X \%*\% beta)
Y0 <- rep(0, n)

pi1 <- 2/3
n1 <- ceiling(n * pi1)
ind <- sample(n, size = n1)
A <- rep(0, n)
A[ind] <- 1
Y <- Y1 * A + Y0 * (1 - A)

Xc_svd <- svd(X)
H <- Xc_svd$u \%*\% t(Xc_svd$u)

result_ls <- esti_mean_treat(X, Y, A, H)
point_est <- result_ls$point_est
var_est <- result_ls$var_est
print(paste0('True mean treat:', round(mean(Y1), digits = 3), '.'))
print('Absolute bias:')
print(abs(point_est - mean(Y1)))
print('Estimate variance:')
print(var_est)

}
\references{
Winston Lin. Agnostic notes on regression adjustments to experimental data: Reexamining Freedman’s critique. The Annals of Statistics, 7(1):295–318, 2013.

Lihua Lei, Peng Ding. Regression adjustment in completely randomized experiments with a diverging number of covariates. Biometrika, 815–828, 2020.

Sihui Zhao, Xinbo Wang, Lin Liu, & Xin Zhang. Covariate adjustment in randomized experiments motivated by higher-order influence functions. arXiv preprint. https://arxiv.org/abs/2411.08491.

Xin Lu, Fan Yang, and Yuhao Wang. Debiased regression adjustment in completely randomized experiments with moderately high-dimensional covariates. arXiv preprint arXiv:2309.02073, 2023.
}
