% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_estimate_mean_treat.R
\name{fit.ate.adj2.adj2c.Random.CRE}
\alias{fit.ate.adj2.adj2c.Random.CRE}
\title{Covariate-Adjusted Treatment Effect Estimation under the Randomization-based Framework}
\usage{
fit.ate.adj2.adj2c.Random.CRE(Y, X, A)
}
\arguments{
\item{Y}{Numeric vector of length n containing observed responses.}

\item{X}{Numeric matrix (n x p) of covariates. Centering is required. May include intercept column.}

\item{A}{Binary vector of length n indicating treatment assignment (1 = treatment, 0 = control).}
}
\value{
A list containing point estimates and variance estimates:
\describe{
   \item{tau_vec}{Point estimates for adj2 and adj2c:
       \itemize{
         \item{\code{adj2}:} Point estimation of the HOIF-inspired debiased estimator (Zhao et al., 2024).
         \item{\code{adj2c}:} Point estimation of the debiased estimator given by Lu et al. (2023), which is also the HOIF-inspired debiased estimator (Zhao et al., 2024)
       }
   }
   \item{var_vec}{Variance estimates for adj2 and adj2c:
     \itemize{
        \item{\code{adj2}:} Variance estimation of adj2, with formulas inspired by Lu et al. (2023).
        \item{\code{adj2c}:} Variance estimation of adj2c, with formulas given by Lu et al. (2023).
        \item{\code{adj2_v2}:} Variance estimation of adj2, with formulas given in Zhao et al. (2024), which is more conservative  compared to \code{adj2}.
        \item{\code{adj2c_v2}:} Variance estimation of adj2c, with formulas given in Zhao et al. (2024), which is more conservative  compared to \code{adj2c}.
     }
   }
}
}
\description{
Implements the debiased estimators for average treatment effect (ATE) estimation
in completely randomized experiments with moderately high-dimensional covariates.
}
\examples{
set.seed(100)
n <- 500
p <- n * 0.3
beta <- runif(p, -1 / sqrt(p), 1 / sqrt(p))

X <- mvtnorm::rmvt(n, sigma = diag(1, p), df = 3)
Y1 <- as.numeric(X \%*\% beta)
Y0 <- as.numeric(X \%*\% beta - 1)

pi1 <- 2/3
n1 <- ceiling(n * pi1)
ind <- sample(n, size = n1)
A <- rep(0, n)
A[ind] <- 1
Y <- Y1 * A + Y0 * (1 - A)

Xc <- cbind(1, scale(X, scale = FALSE))
result.adj2.adj2c.random.ls <- fit.ate.adj2.adj2c.Random.CRE(Y, Xc, A)
point_est <- result.adj2.adj2c.random.ls$tau_vec
var_est <- result.adj2.adj2c.random.ls$var_vec
point_est
var_est
}
\references{
Xin Lu, Fan Yang, and Yuhao Wang. Debiased regression adjustment in completely randomized experiments with moderately high-dimensional covariates. arXiv preprint arXiv:2309.02073, 2023.

Sihui Zhao, Xinbo Wang, Lin Liu, & Xin Zhang. Covariate adjustment in randomized experiments motivated by higher-order influence functions. arXiv preprint. https://arxiv.org/abs/2411.08491.
}
