% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_estimate_mean_treat.R
\name{fit.mean.adj2.adj2c.Super.SR}
\alias{fit.mean.adj2.adj2c.Super.SR}
\title{Covariate-Adjusted Treatment Effect Estimation under the Super-Population Framework}
\usage{
fit.mean.adj2.adj2c.Super.SR(
  Y,
  X,
  A,
  intercept = FALSE,
  Treated = TRUE,
  pi1_hat = NULL
)
}
\arguments{
\item{Y}{Numeric vector of length n containing observed responses.}

\item{X}{Numeric matrix (n x p) of covariates. Centering is required. May include intercept column.}

\item{A}{Binary vector of length n indicating treatment assignment (1 = treatment, 0 = control).}

\item{intercept}{Logical. If TRUE (default), X already contains intercept. Set FALSE if X does not contain intercept.}

\item{Treated}{Logical. If TRUE (default), consider the treatment arm. Set FALSE if considering the control arm.}

\item{pi1_hat}{Default is NULL. The assignment probability for the simple randomization.}
}
\value{
A list containing three named vectors:
\describe{
  \item{tau_vec}{Point estimates:
    \itemize{
      \item{\code{adj2}:} Point estimation of the HOIF-inspired debiased estimator (Zhao et al., 2024).
      \item{\code{adj2c}:} Point estimation of the the HOIF-inspired debiased estimator (Zhao et al., 2024), which is also the debiased estimator given by Lu et al. (2023).
    }}
  \item{var_infl_vec}{Influence function-based variance estimates:
    \itemize{
      \item{\code{adj2}:} Variance for \code{adj2} via the sample variance of its influence function formula.
      \item{\code{adj2c}:} Variance for \code{adj2c} via the sample variance of its influence function formula.
    }}
  \item{var_rb_vec}{Variance estimates inspired by RobinCar:
    \itemize{
      \item{\code{adj2}:} Variance  for \code{adj2} following the asymptotic variance given by Bannick et al. (2025).
      \item{\code{adj2c}:} Variance  for \code{adj2c} following the asymptotic variance given by Bannick et al. (2025).
    }}
}
}
\description{
Implements HOIF-inspired debiased estimators for  treatment effect on the treatment/control arm, with variance estimation
using influence function-based and asymptotic-variance. Designed for simple randomization experiments with moderately high-dimensional covariates.
}
\examples{

set.seed(120)
alpha0 <- 0.1;
n <- 400;

p0 <- ceiling(n * alpha0)
beta0_full <- 1 / (1:p0) ^ (1 / 2) * (-1) ^ c(1:p0)
beta <- beta0_full / norm(beta0_full,type='2')

Sigma_true <- matrix(0, nrow = p0, ncol = p0)
for (i in 1:p0) {
  for (j in 1:p0) {
    Sigma_true[i, j] <- 0.1 ** (abs(i - j))
  }
}

X <- mvtnorm::rmvt(n, sigma = Sigma_true, df = 3)

lp0 <- X \%*\% beta
delta_X <- 1  -  1/4 * X[, 2] -  1/8 * X[, 3]
lp1 <- lp0 + delta_X

Y0 <- lp0 + rnorm(n)
Y1 <- lp1 + rnorm(n)


pi1 <- 1 / 2
A <- rbinom(n, size = 1, prob = pi1)
Y <- A * Y1 + (1 - A) * Y0

Xc <- cbind(1, scale(X, scale = FALSE))
result.treat.adj2.adj2c.sp.ls <- fit.mean.adj2.adj2c.Super.SR(Y, Xc, A, intercept = TRUE, Treated = TRUE)
result.control.adj2.adj2c.sp.ls <- fit.mean.adj2.adj2c.Super.SR(Y, Xc, A, intercept = TRUE, Treated = FALSE)
result.treat.adj2.adj2c.sp.ls
result.control.adj2.adj2c.sp.ls

}
\references{
Marlena S. Bannick, Jun Shao, Jingyi Liu, Yu Du, Yanyao Yi, Ting Ye (2025). A General Form of Covariate Adjustment in Randomized Clinical Trials. Biometrika.

Xin Lu, Fan Yang, and Yuhao Wang. Debiased regression adjustment in completely randomized experiments with moderately high-dimensional covariates. arXiv preprint arXiv:2309.02073, 2023.

Sihui Zhao, Xinbo Wang, Lin Liu, & Xin Zhang. Covariate adjustment in randomized experiments motivated by higher-order influence functions. arXiv preprint. https://arxiv.org/abs/2411.08491.
}
