% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_oracle_infer.R
\name{get_oracle_bias_var_adj2_dag}
\alias{get_oracle_bias_var_adj2_dag}
\title{Estimate the oracle bias, the exact variance and approximated variance of the debiased estimator motivated by HOIF.}
\usage{
get_oracle_bias_var_adj2_dag(X, Y1, pi1 = 0.5)
}
\arguments{
\item{X}{The n by p covariates matrix.}

\item{Y1}{Vector of n dimensional potential response Y(1).}

\item{pi1}{The proportion of subjects in the treatment group.}
}
\value{
A list of oracle bias and variance.
\item{bias_adj2_dag}{The oracle bias of the estimator.}
\item{variance_exact_adj2_dag}{The oracle exact variance of the estimator.}
\item{variance_approx_adj2_dag}{The oracle approximated variance of the estimator which omits the term of order o(1/n).}
}
\description{
Estimate the oracle bias, the exact variance and approximated variance of the debiased estimator motivated by HOIF.
}
\references{
Zhao, S., Wang, X., Liu, L., & Zhang, X. (2024). Covariate adjustment in randomized experiments motivated by higher-order influence functions. arXiv preprint. https://arxiv.org/abs/2411.08491
}
