% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_oracle_infer.R
\name{get_oracle_var_adj_db}
\alias{get_oracle_var_adj_db}
\title{Estimate the oracle bias, the oracle variance of the unadjusted estimator, the Lin’s (2013) adjusted estimator and the debiased estimator tau_db by Lu et al.(2023).}
\usage{
get_oracle_var_adj_db(X, Y1, pi1 = 0.5)
}
\arguments{
\item{X}{The n by p covariates matrix.}

\item{Y1}{Vector of n dimensional potential response Y(1).}

\item{pi1}{The proportion of subjects in the treatment group.}
}
\value{
A list of variance_unadj, variance_adj and variance_db.
\item{variance_unadj}{The oracle variance of the unadjusted estimator.}
\item{variance_adj}{The oracle variance of Lin’s (2013) adjusted estimator tau_adj with linear working model.}
\item{variance_db}{The oracle variance of the debiased estimator tau_db by Lu et al.(2023).}
}
\description{
Estimate the oracle bias, the oracle variance of the unadjusted estimator, the Lin’s (2013) adjusted estimator and the debiased estimator tau_db by Lu et al.(2023).
}
\references{
Winston Lin. Agnostic notes on regression adjustments to experimental data: Reexamining Freedman’s critique. The Annals of Statistics, 7(1):295–318, 2013.

Xin Lu, Fan Yang, and Yuhao Wang. Debiased regression adjustment in completely randomized experiments with moderately high-dimensional covariates. arXiv preprint arXiv:2309.02073, 2023.
}
