% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mean_treat.R
\name{esti_mean_treat}
\alias{esti_mean_treat}
\title{Estimate treatment effect on the treatment arm using different covariates adjustment methods.}
\usage{
esti_mean_treat(X, Y, A, H = NULL)
}
\arguments{
\item{X}{The n by p covariates matrix}

\item{Y}{Vector of n dimensional observed response.}

\item{A}{Vector of n dimensional treatment assignment.}

\item{H}{The n by n hat matrix corresponding to X.}
}
\value{
\item{tau_unadj}{The point estimation of unadjusted estimator tau_unadj.}
\item{tau_adj}{The point estimation of Lin’s (2013) adjusted estimator with linear working model.}
\item{tau_db}{The point estimation of debiased estimator tau_db by Lu et al.(2023).}
\item{tau_adj2}{The point estimation of HOIF-motivated adjusted estimator.}
\item{tau_adj2_dag}{The point estimation of another form of debiased estimator tau_db.}
\item{tau_adj3}{The point estimation of bias-free adjusted estimator based on tau_adj2.}
}
\description{
Estimate treatment effect on the treatment arm using different covariates adjustment methods.
}
\examples{
set.seed(100)
n <- 500
p <- n * 0.3
beta <- runif(p, -1 / sqrt(p), 1 / sqrt(p))

X <- mvtnorm::rmvt(n, sigma = diag(1, p), df = 2)
Y1 <- as.numeric(X \%*\% beta)
Y0 <- rep(0, n)

pi1 <- 0.45
n1 <- n * pi1
ind <- sample(n, size = n1)
A <- rep(0, n)
A[ind] <- 1
Y <- Y1 * A + Y0 * (1 - A)

Xc_svd <- svd(X)
H <- Xc_svd$u \%*\% t(Xc_svd$u)

result_vec <- esti_mean_treat(X, Y, A, H)
print(paste0('True mean treat:', round(mean(Y1), digits = 3), '.'))
print('Absolute bias:')
print(abs(result_vec - mean(Y1)))


}
\references{
Winston Lin. Agnostic notes on regression adjustments to experimental data: Reexamining Freedman’s critique. The Annals of Statistics, 7(1):295–318, 2013.

Xin Lu, Fan Yang, and Yuhao Wang. Debiased regression adjustment in completely randomized experiments with moderately high-dimensional covariates. arXiv preprint arXiv:2309.02073, 2023.
}
